Attempting to fix circular imports by detatching some typing stuff so that classifiers and projects can both import them
