+++ src/classifier/__init__.py
@@ -1,2 +1,1 @@
-from .classifier import ClassifierType
 from .classifier import Classifier
