+++ src/project/__init__.py
@@ -1,5 +1,5 @@
 from .track_labels import TrackLabels
 from .video_labels import VideoLabels
 from .project import Project
-from .units import ProjectDistanceUnit
-from .export_training import export_training_data, load_training_data
+from .export_training import export_training_data
+from .read_training import load_training_data
