+++ src/types/classifier_types.py
@@ -1,0 +1,7 @@
+from enum import IntEnum
+
+
+class ClassifierType(IntEnum):
+    RANDOM_FOREST = 1
+    GRADIENT_BOOSTING = 2
+    XGBOOST = 3
